Fix all 37 compilation errors in generated bindings

+++ Metal.NET/Metal/MTLTextureSwizzleChannels.cs
@@ -38,7 +38,7 @@ public class MTLTextureSwizzleChannels : IDisposable
 
     public static MTLTextureSwizzleChannels Make(MTLTextureSwizzle r, MTLTextureSwizzle g, MTLTextureSwizzle b, MTLTextureSwizzle a)
     {
-        MTLTextureSwizzleChannels result = new(ObjectiveCRuntime.MsgSendPtr(Class, MTLTextureSwizzleChannelsSelector.MakeGBA, (ulong)r, (ulong)g, (ulong)b, (ulong)a));
+        MTLTextureSwizzleChannels result = new(ObjectiveCRuntime.MsgSendPtr(Class, MTLTextureSwizzleChannelsSelector.MakeGBA, (byte)r, (byte)g, (byte)b, (byte)a));
 
         return result;
     }
